transfer io_utils to meta_utils
